+++ src/protocols/minecraft/protocol/bedrock.rs
@@ -83,7 +83,7 @@ impl Bedrock {
             edition: status[0].to_string(),
             name: status[1].to_string(),
             version_name: status[3].to_string(),
-            version_protocol: status[2].to_string(),
+            protocol_version: status[2].to_string(),
             players_maximum: status[5].parse().map_err(|e| TypeParse.context(e))?,
             players_online: status[4].parse().map_err(|e| TypeParse.context(e))?,
             id: status.get(6).map(|v| v.to_string()),