+++ src/protocols/minecraft/protocol/legacy_v1_6.rs
@@ -52,11 +52,11 @@ impl LegacyV1_6 {
 
     pub(crate) fn get_response(buffer: &mut Buffer<BigEndian>) -> GDResult<JavaResponse> {
         // This is a specific order!
-        let version_protocol = buffer
+        let protocol_version = buffer
             .read_string::<Utf16Decoder<BigEndian>>(None)?
             .parse()
             .map_err(|e| PacketBad.context(e))?;
-        let version_name = buffer.read_string::<Utf16Decoder<BigEndian>>(None)?;
+        let game_version = buffer.read_string::<Utf16Decoder<BigEndian>>(None)?;
         let description = buffer.read_string::<Utf16Decoder<BigEndian>>(None)?;
         let online_players = buffer
             .read_string::<Utf16Decoder<BigEndian>>(None)?
@@ -68,11 +68,11 @@ impl LegacyV1_6 {
             .map_err(|e| PacketBad.context(e))?;
 
         Ok(JavaResponse {
-            version_name,
-            version_protocol,
+            game_version,
+            protocol_version,
             players_maximum: max_players,
             players_online: online_players,
-            players_sample: None,
+            players: None,
             description,
             favicon: None,
             previews_chat: None,