+++ src/protocols/quake/two.rs
@@ -32,7 +32,7 @@ impl CommonPlayer for Player {
 
     fn name(&self) -> &str { &self.name }
 
-    fn score(&self) -> Option<u32> { Some(self.score.try_into().unwrap_or(0)) }
+    fn score(&self) -> Option<i32> { Some(self.score) }
 }
 
 pub(crate) struct QuakeTwo;